+++ workers/ohsome_quality_analyst/base/indicator.py
@@ -13,11 +13,7 @@ import matplotlib.pyplot as plt
 from dacite import from_dict
 from geojson import Feature
 
-from ohsome_quality_analyst.utils.definitions import (
-    INDICATOR_LAYER,
-    get_layer_definition,
-    get_metadata,
-)
+from ohsome_quality_analyst.utils.definitions import get_layer_definition, get_metadata
 
 
 @dataclass
@@ -83,8 +79,6 @@ class BaseIndicator(metaclass=ABCMeta):
             svg=self._get_default_figure(),
         )
 
-        self._validate_indicator_layer(self.__class__.__name__, layer_name)
-
     def as_feature(self) -> Feature:
         """Returns a GeoJSON Feature object.
 
@@ -145,13 +139,6 @@ class BaseIndicator(metaclass=ABCMeta):
         Writes an SVG figure to the svg attribute of the result attribute.
         """
         pass
-
-    def _validate_indicator_layer(self, indicator_name, layer_name):
-        indicator_layer = (indicator_name, layer_name)
-        if indicator_layer not in INDICATOR_LAYER:
-            raise ValueError(
-                "Indicator layer combination is invalid: " + str(indicator_layer)
-            )
 
     def _get_default_figure(self) -> str:
         """Return a SVG as default figure for indicators"""