+++ workers/tests/integrationtests/test_api_indicator.py
@@ -5,6 +5,7 @@ https://fastapi.tiangolo.com/tutorial/testing/
 import os
 import unittest
 from typing import Optional
+from urllib.parse import urlencode
 
 from fastapi.testclient import TestClient
 from schema import Schema
@@ -13,6 +14,8 @@ from ohsome_quality_analyst.api.api import app
 
 from .api_response_schema import get_general_schema, get_indicator_feature_schema
 from .utils import oqt_vcr
+
+ENDPOINT = "/indicator"
 
 
 class TestApiIndicator(unittest.TestCase):
@@ -31,6 +34,7 @@ class TestApiIndicator(unittest.TestCase):
 
     def run_tests(self, response) -> None:
         self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.headers["content-type"], "application/geo+json")
         for schema in (self.general_schema, self.feature_schema):
             self.validate(response.json(), schema)
 
@@ -47,23 +51,15 @@ class TestApiIndicator(unittest.TestCase):
         fid_field: Optional[str] = None,
     ):
         """Return HTTP GET response"""
+        parameters_raw = {
+            "name": indicator_name,
+            "layerName": layer_name,
+            "dataset": dataset,
+            "featureId": feature_id,
+        }
         if fid_field is not None:
-            base_url = "/indicator/{0}?".format(self.indicator_name)
-            parameter = "layerName={0}&dataset={1}&featureId={2}&fidField={3}".format(
-                layer_name,
-                dataset,
-                feature_id,
-                fid_field,
-            )
-            url = base_url + parameter
-        else:
-            url = "/indicator/{0}?layerName={1}&dataset={2}&featureId={3}".format(
-                indicator_name,
-                layer_name,
-                dataset,
-                feature_id,
-            )
-        return self.client.get(url)
+            parameters_raw["fidField"] = fid_field
+        return self.client.get(ENDPOINT + "?" + urlencode(parameters_raw))
 
     def post_response(
         self,
@@ -76,19 +72,20 @@ class TestApiIndicator(unittest.TestCase):
         """Return HTTP POST response"""
         if fid_field is not None:
             data = {
-                "dataset": self.dataset,
-                "featureId": self.feature_id,
-                "layerName": self.layer_name,
+                "name": indicator_name,
+                "dataset": dataset,
+                "layerName": layer_name,
+                "featureId": feature_id,
+                "fidField": fid_field,
             }
         else:
             data = {
+                "name": indicator_name,
                 "dataset": dataset,
-                "featureId": feature_id,
                 "layerName": layer_name,
-                "fidField": fid_field,
+                "featureId": feature_id,
             }
-        url = "/indicator/{0}".format(indicator_name)
-        return self.client.post(url, json=data)
+        return self.client.post(ENDPOINT, json=data)
 
     @oqt_vcr.use_cassette()
     def test_indicator_dataset_default_fid_field(self):
@@ -137,11 +134,12 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_dataset_invalid(self):
         data = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
             "dataset": "foo",
             "featureId": "3",
         }
-        url = "/indicator/{0}".format(self.indicator_name)
-        response = self.client.post(url, json=data)
+        response = self.client.post(ENDPOINT, json=data)
         self.assertEqual(response.status_code, 422)
 
     @oqt_vcr.use_cassette()
@@ -153,17 +151,20 @@ class TestApiIndicator(unittest.TestCase):
         )
         with open(path, "r") as f:
             bpolys = f.read()
-        url = "/indicator/{0}".format(self.indicator_name)
         for data in (
             {
+                "name": self.indicator_name,
                 "bpolys": bpolys,
                 "dataset": "foo",
                 "featureId": "3",
             },
-            {"dataset": "regions"},
-            {"feature_id": "3"},
+            {
+                "name": self.indicator_name,
+                "dataset": "regions",
+            },
+            {"name": self.indicator_name, "feature_id": "3"},
         ):
-            response = self.client.post(url, json=data)
+            response = self.client.post(ENDPOINT, json=data)
             self.assertEqual(response.status_code, 422)
 
     @oqt_vcr.use_cassette()
@@ -197,6 +198,17 @@ class TestApiIndicator(unittest.TestCase):
         response = self.client.get(url)
         result = response.json()
         self.assertNotIn("result.svg", list(result["properties"].keys()))
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_invalid_layer(self):
+        data = {
+            "name": self.indicator_name,
+            "layerName": "amenities",
+            "dataset": "regions",
+            "featureId": "3",
+        }
+        response = self.client.post(ENDPOINT, json=data)
+        self.assertEqual(response.status_code, 422)
 
 
 if __name__ == "__main__":