+++ workers/tests/integrationtests/test_api_indicator_geojson_io.py
@@ -6,7 +6,9 @@ https://fastapi.tiangolo.com/tutorial/testing/
 import os
 import unittest
 from typing import Tuple
+from urllib.parse import urlencode
 
+import geojson
 from fastapi.testclient import TestClient
 from schema import Schema
 
@@ -23,7 +25,7 @@ from .utils import oqt_vcr
 def get_fixture(name):
     path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "fixtures", name)
     with open(path, "r") as f:
-        return f.read()
+        return geojson.load(f)
 
 
 class TestApiIndicatorIo(unittest.TestCase):
@@ -39,6 +41,7 @@ class TestApiIndicatorIo(unittest.TestCase):
 
     def run_tests(self, response, schemata: Tuple[Schema]) -> None:
         self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.headers["content-type"], "application/geo+json")
         for schema in schemata:
             self.validate(response.json(), schema)
 
@@ -48,17 +51,24 @@ class TestApiIndicatorIo(unittest.TestCase):
 
     def get_response(self, bpoly):
         """Return HTTP GET response"""
-        url = "/indicator/{0}?layerName={1}&bpolys={2}".format(
-            self.indicator_name,
-            self.layer_name,
-            bpoly,
+        parameters = urlencode(
+            {
+                "name": self.indicator_name,
+                "layerName": self.layer_name,
+                "bpolys": bpoly,
+            }
         )
+        url = "/indicator?" + parameters
         return self.client.get(url)
 
     def post_response(self, bpoly):
         """Return HTTP POST response"""
-        data = {"bpolys": bpoly, "layerName": self.layer_name}
-        url = f"/indicator/{self.indicator_name}"
+        data = {
+            "name": self.indicator_name,
+            "bpolys": bpoly,
+            "layerName": self.layer_name,
+        }
+        url = "/indicator"
         return self.client.post(url, json=data)
 
     @oqt_vcr.use_cassette()
@@ -106,8 +116,12 @@ class TestApiIndicatorIo(unittest.TestCase):
         )
         with open(path, "r") as file:
             bpolys = file.read()
-        data = {"bpolys": bpolys, "layerName": self.layer_name}
-        url = f"/indicator/{self.indicator_name}"
+        data = {
+            "name": self.indicator_name,
+            "bpolys": bpolys,
+            "layerName": self.layer_name,
+        }
+        url = "/indicator"
         response = self.client.post(url, json=data)
         self.assertEqual(response.status_code, 422)