+++ CHANGELOG.md
@@ -1,10 +1,28 @@
 # Changelog
 
-## 0.7.0
+## Current Main
 
+### Breaking Changes
+
+- Change API Path parameter `name` to be a Query parameter instead of a Path parameter ([#190])
+- Change type of API parameter `bpolys` for POST requests to JSON (`dict`) instead of string ([#204])
+
+### Other Changes
+
+- API: Change media type of GeoJSON response ([#199])
+- Validate indicator-layer combination with `pydantic` ([#190])
 - Add parameter to get API response without svg string in the result. ([#137])
 
+### How to upgrade
+
+- Update requests for `/indicator` and `/report` endpoints. `name` is now a *Query* parameter ([#190])
+    - E.g. `/indicator/GhsPopComparisonBuilding` -> `indicator?name=GhsPopComparisonBuilding`
+- Update `bpolys` parameter of POST requests to be a GeoJSON object instead of a string ([#204])
+
 [#137]: https://github.com/GIScience/ohsome-quality-analyst/issues/137
+[#190]: https://github.com/GIScience/ohsome-quality-analyst/pull/190
+[#199]: https://github.com/GIScience/ohsome-quality-analyst/pull/199
+[#204]: https://github.com/GIScience/ohsome-quality-analyst/pull/204
 
 
 ## 0.6.0